separate resource from controller due to tree shaking issues in dev
