+++ api/src/Blog/FindPost.ts
@@ -1,18 +1,8 @@
 import { handlerFor } from "#api/lib/handler"
-import { S } from "#resources/lib"
 import { Effect } from "effect"
 import { Option } from "effect-app"
-import { BlogPostId } from "./models.js"
+import { FindPost } from "./FindPost.resource.js"
 import { BlogPostRepo } from "./Repo.js"
-import { BlogPostView } from "./views.js"
-
-export class FindPost extends S.Req<FindPost>()("Blog.FindPost", {
-  id: BlogPostId
-}, {
-  allowAnonymous: true,
-  allowRoles: ["user"],
-  success: S.NullOr(BlogPostView)
-}) {}
 
 export default handlerFor(FindPost)({
   dependencies: [
