+++ api/src/Blog/ListPosts.ts
@@ -1,16 +1,7 @@
 import { handlerFor } from "#api/lib/handler"
-import { S } from "#resources/lib"
 import { Effect } from "effect-app"
+import { ListPosts } from "./ListPosts.resource.js"
 import { BlogPostRepo } from "./Repo.js"
-import { BlogPostView } from "./views.js"
-
-export class ListPosts extends S.Req<ListPosts>()("Blog.List", {}, {
-  allowAnonymous: true,
-  allowRoles: ["user"],
-  success: S.Struct({
-    items: S.Array(BlogPostView)
-  })
-}) {}
 
 export default handlerFor(ListPosts)({
   dependencies: [
