+++ api/src/Blog/PublishPost.ts
@@ -2,17 +2,10 @@ import { handlerFor } from "#api/lib/handler"
 import { OperationsDefault } from "#api/lib/layers"
 import { Events, Operations } from "#api/services"
 import { BogusEvent } from "#resources/Events"
-import { S } from "#resources/lib"
 import { Duration, Effect, Schedule } from "effect-app"
-import { NotFoundError } from "effect-app/client"
-import { OperationId } from "effect-app/Operations"
 import { NonEmptyString2k, NonNegativeInt } from "effect-app/Schema"
-import { BlogPostId } from "./models.js"
+import { PublishPost } from "./PublishPost.resource.js"
 import { BlogPostRepo } from "./Repo.js"
-
-export class PublishPost extends S.Req<PublishPost>()("Blog.PublishPost", {
-  id: BlogPostId
-}, { allowRoles: ["user"], success: OperationId, failure: S.Union(NotFoundError) }) {}
 
 export default handlerFor(PublishPost)({
   dependencies: [
