+++ api/src/Operations/FindOperation.ts
@@ -1,13 +1,8 @@
 import { handlerFor } from "#api/lib/handler"
 import { OperationsDefault } from "#api/lib/layers"
 import { Operations } from "#api/services"
-import { S } from "#resources/lib"
 import { Effect } from "effect-app"
-import { Operation, OperationId } from "effect-app/Operations"
-
-export class FindOperation extends S.Req<FindOperation>()("Operations.FindOperation", {
-  id: OperationId
-}, { allowAnonymous: true, allowRoles: ["user"], success: S.NullOr(Operation) }) {}
+import { FindOperation } from "./FindOperation.resource.js"
 
 export default handlerFor(FindOperation)({
   dependencies: [OperationsDefault],
