+++ api/src/resources/HelloWorld.ts
@@ -1,4 +1,4 @@
-export { GetHelloWorld } from "../HelloWorld/GetHelloWorld.js"
+export { GetHelloWorld } from "../HelloWorld/GetHelloWorld.resource.js"
 
 // codegen:start {preset: meta, sourcePrefix: src/resources/}
 export const meta = { moduleName: "HelloWorld" } as const
